requirements.txt
pandas==1.0.5
numpy==1.19.5
numba==0.53.1
scipy==1.5.0
torch==1.4.0
python==3.6.5
gensim==3.8.0
deepctr-torch==0.2.7
transformers==3.1.0
bayesian-optimization==1.2.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0